--- src/components/general/index.ts
@@ -1,12 +1,0 @@
-export * from './AccordionList';
-export * from './BottomSheet';
-export * from './CodeBlock';
-export * from './Collapsible';
-export * from './Fab';
-export * from './Headings';
-export * from './IconButton';
-export * from './Input';
-export * from './List';
-export * from './Radio';
-export * from './Separator';
-export * from './Tooltip';
