+++ src/components/layout/BaseColorsSection/ColorListItem.tsx
@@ -1,8 +1,8 @@
 import {
   AccordionListItem,
-  ListItemText,
   type AccordionListItemProps,
-} from '@/components/general';
+} from '@/components/ui/AccordionList';
+import { ListItemText } from '@/components/ui/List';
 
 interface ColorListItemProps extends AccordionListItemProps {
   value: string;
